test: create fixtures handler
